migration hpp to h and cpp
